Pass the "available modules" for a target down to Swift commands
When the Swift driver can handle it, produce the complete set of modules
that are available within a package (either from targets in that package or
products of packages it depends on). Provide that set of available modules
as a JSON file to the Swift driver so it can reason about modules that
are available to the given target. Also provide the package manifest
path and the current target name in the package for use by the compiler.