Generated code for XCTest on non-Darwin needs to be actor-isolated. (#7566)
On Linux, Windows, etc. (anywhere that uses swift-corelibs-xctest
instead of XCTest.framework), SwiftPM is responsible for generating an
entry point function that passes in all tests discovered at compile
time. The compiler cannot tell whether the generated code is
concurrency-safe. This PR modifies the generated code and makes it
main-actor-isolated. Since it's only ever used in the program's main
function, this is safe.

Resolves #7556.