[Build/Commands/PackageGraph] Remove `ModulesGraph.updateBuildTripleRecursively`
Plugins don't actually need whole graph to be switched to use
`.tools` triple because `PluginTarget.processAccessibleTools`
is able to find targets/products that have to be built for the
host, it just needs a way to signal that to the build system.

`Buildset` has been extended to support a destination which
is defaulted to `.target` because really only plugin (command) tools
have to set it to `.host`, this helps to propagate the information
about expected destination down through `BuildOperation.build(subset:)`
and find correct llbuild names and build products.