[Plugins] Make sure that plugins work when referenced by tools and destination modules
Remove incorrect assumption that the outputs of the plugin are always
connected to a destination build of a particular module and add a build
triple into the path to separate the outputs of a plugin.

This comes into play when i.e. a test is built for both `tools` and
`destination`, we need two plugin invocations to handle that.