Add cxx interop support to symbolgraph-extract
In order to extract symbol graphs for modules with transitive imports on
cxx modules we need parse headers in cxx mode using the option:
`-cxx-interoperability-mode=default`. This commit updates
`swift-symbolgraph-extract` to pass the option when there is a Cxx
module in the import graph.

This option is a new addition to `swift-symbolgraph-extract` added in
apple/swift#73963.

Prior to this option's introduction, extracting the symbol graph for cxx
dependent modules would fail with an error similar to:
"unknown type name 'namespace'".

With this SwiftPM commit and `swift-symbolgraph-extract` _prior_ to
73963, users will instead see an argument parsing error like: "unknown
option -cxx-interoperability-mode".

With this change and `swift-symbolgraph-extract` 73963, the symbol graph
is extracted properly.