Fix `swift-api-digester` search paths
Fixes an architecture and implementation bug where the include paths
for `swift-api-digester` were being determined inside build plan instead
of by the target build description.

This is effectively the same change as #7621 but for api-digester rather
than symbolgraph-extract.

This commit also leaves hooks in to enable api-digester on clang modules
as a future improvement.