[Build] Adjust `computeLLBuildTargetName` to attempt `host` destination for test products
Just like macros and plugins, tests products can also be built for
the host depending on their test target(s) destinations (which are
inferred based on dependencies). In order to properly support
`--test-product` option we need to fallback to `host` lookup for
test products just like we do for macros and plugins.