Add cxx interop support to symbolgraph-extract (#7610)
In order to extract symbol graphs for modules with transitive imports on
C++ modules we need parse headers in C++ interop mode using the option:
`-cxx-interoperability-mode=default`. This commit updates
`swift-symbolgraph-extract` to pass the option when there is a C++
module in the import graph.

This option is a new addition to `swift-symbolgraph-extract` added in
apple/swift#73963.

Prior to this option's introduction, extracting the symbol graph for C++
dependent modules would fail with an error similar to: "unknown type
name 'namespace'".

With this SwiftPM commit and `swift-symbolgraph-extract` _prior_ to
apple/swift#73963, users will instead see an argument parsing error
like: "unknown option -cxx-interoperability-mode".

With this change and `swift-symbolgraph-extract` change in
apple/swift#73963, the symbol graph is extracted properly.