[Build/PackageGraph] Switch `BuildSubset.{product, target}` and `Modu… (#7650)
…lesGraph.{product, target}(...)` to use optional destination

This is mostly an NFC change. It makes more sense to default
`destination` to "undefined" when applicable because it's not always
possible to know intended destination based on user input.

### Motivation:

This kind of design makes more sense for internal APIs, instead of
having users to pass `.destination` even though it might not be always
correct.

### Modifications:

- `BuildSubset.{product, target}` and `ModulesGraph.{product,
target}(...)` accept an optional `destination` that is defaulted to
`.none`

### Result:

Easier to understand and use APIs to request a build subset and product
and module references from the modules graph.