[Build] Extract build of plugin tools from `invokeBuildToolPlugins` into a separate method on `BuildOperation`
Simplifies `invokeBuildToolPlugins`, removes the need to have a separate
location for plugin tools build and use a separate `BuildOperation`.

Next step would be to factor plugin tool building into a separate operation
which gets run as part of the build and attempt to share `build.db` between
regular build and plugin tools build.