Relax version ranges for dependencies in `Package.swift` (#7655)
This replaces #7649 (which proposed the same change on the 5.10 branch,
I'l open a PR to cherry-pick back to 5.10 and 6.0 as well)

Referencing the rationale from the original PR:

We're currently being held back from pulling in Vapor and NIO updates
due to tighter dependency clauses on swift-system and swift-certificates
in SwiftPM.

We need to depend on SwiftPM in order to use the package collection
signing module. We do this by depending on `revision:
"swift-5.10.1-RELEASE"` in the SwiftPackageIndex server project.