Fix bearer auth when user is token
When logging in to a package registry bearer auth works.
However when fetching binaryTargets or pulling packages from a registry
basic auth is used always.

So the header will basically be "Basic token:acces_token"

This either breaks services that don't use basic auth or leads to nasty
workarounds like first creating a netrc file with "token" as the user
and after login changing the user back to the actual user. Assuming that
basic auth is allowed with an identiy token.

When using "bearer" auth, when the user is "token" these issues are
fixed.