Fix cache not cleaned up if download fails
If the download of a binaryTarget fails a cache file still remains in
the artifact cache.

The file only contains the server's response which is usually the status
code and the error message.

All following resolve calls therefore fail because the cached file is
used but is an invalid archive