Vendor `AsyncProcess`, make more tests `async`
We'd like to implement more changes to `TSCBasic.Process` without fearing of breaking clients, in addition to the fact that TSC is deprecated.

One thing in particular we need is backpressure logic when reading from process instances in plugins communication. Currently, without this logic a naive rewrite of plugins communication with structured concurrency is quite bug-prone.

Thus `TSCBasic.Process` is slightly cleaned up and is vendored as `AsyncProcess` in this change. We also give it more coverage in tests by utilizing `async` overloads, which are the overloads we should be using more in the structured concurrency conversion of the SwiftPM codebase.