[Traits] Integrate traits into module loading (#7689)
# Motivation
In my previous PR I landed the APIs for package traits. We now have to
take this information into consideration when evaluating the module
graph.

# Modification
This PR implements the various trait features in the module graph
loading stage. Including resolving optional dependencies, enabled traits
of dependencies, build settings based on traits and passing defines for
traits.

This PR also contains an exhaustive test fixtures which we can use to
check if all trait functionality is working as expected.