[NFC] Vendor `AsyncProcess`, make more tests `async` (#7679)
We'd like to implement more changes to `TSCBasic.Process` without
fearing of breaking clients, in addition to the fact that TSC is
deprecated.

One thing in particular we need is backpressure logic (coming in a
future PR) when reading from process instances in plugins communication
with the host SwiftPM process. Currently without this logic, a naive
rewrite of plugins communication with structured concurrency is quite
bug-prone, as it loses messages ordering and can quickly fill up all
memory with resource-intensive plugins.

Thus `TSCBasic.Process` is slightly cleaned up and is vendored as
`AsyncProcess` in this change. We also give it more coverage in tests by
utilizing `async` overloads, which are the overloads we should be using
more in the structured concurrency conversion of the SwiftPM codebase.

(cherry picked from commit b86d22e0fac209845982d625a4fd478b037433ca)

# Conflicts:
#	Tests/CommandsTests/PackageCommandTests.swift
#	Tests/FunctionalTests/MiscellaneousTests.swift
#	Tests/FunctionalTests/PluginTests.swift
#	Tests/FunctionalTests/TraitTests.swift