Implement `fileAffectsSwiftOrClangBuildSettings` with the logic from `LLBuildManifestBuilder`
Instead of inspecting file extensions to decide whether a file might affect build settings, check which files are in directories that are affecting compilation. This is the same logic that `LLBuildManifestBuilder` uses and should be more stable. In particular, this stops us from reloading the package manifest in SourceKit-LSP when a header file is written to a `.build` directory.

rdar://128573306