[Traits] Enable trait configuration for `swift-test`
# Motivation

It is useful to be able to configure the traits of the root package when running tests.

# Modification

This PR adds the trait CLI options to the `swift test` command to enable specific traits or all traits, or to disable the default traits.

# Result

`swift test` now supports configuring the root traits.