[Traits] Correctly diagnose unused dependency with traits (#7702)
# Motivation

We are currently diagnosing unused dependencies by looking at the list
of products used by the final modules and check if for every root
dependency we are using at least one product of that package. This
incorrectly diagnosis unused dependencies if the usage of the dependency
is either:
- Just declared to enable a trait if down the graph another package is
depending on it
- Used as an optional dependency and the trait is disabled

# Modification

This PR adds two more checks to the diagnostic pass to find unused
dependencies. We first create a set of dependency products that are
guarded by a trait. When checking each dependency we then check if we
enabled any trait on the dependency since this means it is used for
setting a trait. We also check if any product of the dependency is
guarded by a trait to identify if it was an optional dependency.

# Result
No more incorrect unused dependency warnings.