[6.0] Implement `fileAffectsSwiftOrClangBuildSettings` with the logic from `LLBuildManifestBuilder` (#7709)
- **Explanation**: Instead of inspecting file extensions to decide
whether a file might affect build settings, check which files are in
directories that are affecting compilation. This is the same logic that
`LLBuildManifestBuilder` uses and should be more stable. In particular,
this stops us from reloading the package manifest in SourceKit-LSP when
a header file is written to a `.build` directory.
- **Scope**: Only affects SourceKit-LSP
- **Risk**: Low, only affects SourceKit-LSP
- **Testing**: Adding a SourceKit-LSP test case in
https://github.com/swiftlang/sourcekit-lsp/pull/1512
- **Issue**: rdar://128573306
- **Reviewer**: @MaxDesiatov @xedin on
https://github.com/swiftlang/swift-package-manager/pull/7699