[NFC] Rename `SPMTestSupport` to `_InternalTestSupport` (#7676)
There's no such product or entity as "SPM", this contraction is just as
invalid as "MOS" would be for macOS, or "SNIO" for SwiftNIO. All
marketing and documentation materials consistently refer to it as
"SwiftPM", and we should stay consistent in our codebase too. This also
simplifies code search, as anything named as "SPM" almost never shows up
in results when searching globally for "SwiftPM".

We're not using `SwiftPMTestSupport` name here, as it might mislead
contributors into thinking this module is somehow related to the `swift
test` command. This module is purely internal, used only in modules with
unit tests for SwiftPM itself, and is not supposed to be used by any
packages depending on SwiftPM. Thus all `@_spi` declarations in it were
also changed to `package`.