Fix checksum computation for Swift SDK bundles (#7748)
This feature [was specified in the corresponding proposal for Swift
SDKs](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0387-cross-compilation-destinations.md#swift-sdk-installation-and-configuration):

> For Swift SDKs installed from remote URLs an additional `--checksum`
option is required, through which users of a Swift SDK can specify a
checksum provided by a publisher of the SDK. The latter can produce a
checksum by running `swift package compute-checksum` command (introduced
in
[SE-0272](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0272-swiftpm-binary-dependencies.md))
with the Swift SDK bundle archive as an argument.

Currently, `swift package compute-checksum` is unable to handle
`.tar.gz` bundles, which is a commonly used format for Swift SDKs. We're
fixing that here by adding correct and unified handling for archive
extensions.

This change is isolated to the `swift package compute-checksum`
subcommand. The `--checksum` option is added in a subsequent PR:
https://github.com/swiftlang/swift-package-manager/pull/7722.