Build one test product for both Swift Testing and XCTest (with Swift Testing enabled by default.) (#7766)
This PR refactors the previously-experimental Swift Testing support
logic so that only a single build product is produced when using both
XCTest and Swift Testing, and detection of Swift Testing usage is no
longer needed at compile time.

On macOS, Xcode 16 is responsible for hosting Swift Testing content, so
additional changes may be needed in Xcode to support this refactoring.
Such changes are beyond the purview of the Swift open source project.

Resolves rdar://120864035.

---------

Co-authored-by: Pavel Yaskevich <xedin@apache.org>