Suppress Swift Testing invocation when XCTMain.swift or LinuxMain.swift is present.
If a test entry point source file is present in a package, that means our own
entry point code won't run, so we can't reliably run Swift Testing. This is a
legacy configuration but is still used by a number of packages in the Swift
ecosystem.

Further deprecation of this feature is something to consider after Swift 6.0.