Add destination directory as argument to `swift package init`
Closes #7791

This allows for specifying the directory to create a new package within. If no argument is supplied, matching the current supported command options, then the package will be created in the current working directory. If an argument is supplied, then the package will be created in the directory at the path specified.