Fix Package.swift diagnostic line off by 1
SPM was implicitly preprending the manifest with 'import Foundation' for
toolchains using swift version 5.7 and under. The original file itself
was never overwritten, thus the diagnostics being reported would always
present on the next line.

This is a quick fix that appends the line to the end of the file;
doing so assures that the intended behaviour doesn't change, but allows
for correct diagnostics to be reported.