Ensure code coverage is output correctly when `--parallel` is specified.
The documentation for "LLVM_PROFILE_FILE" suggests, but does not state outright,
that `%m` is unreliable when used with concurrently running processes. This PR
changes the paths we use for code coverage .profraw files to be unique per
invocation using `UUID`. That means that when `--parallel` is specified, the
various processes don't stomp on each other's output.

Resolves #7069.