Move the 'GitRepositoryProvider' to compare urls using their canonical representation, correctly accepting results where repositories only differ by '.git' (#7741)
Resolving local package dependencies often outputs a warning:
<path> is not valid git repository for '<repo>', will fetch again.

The underlying issue is that updating packages with a local SCM path
dependency always fail, as the shell-based repository URL has a file://
scheme, while the absolute path string does not.

After this change we only compare the canonical URLs, which will resolve
the inconsistency of comparing strings without considering the scheme
and / or path extension.

Validating the origin of checked-out repositories no longer ensures that
the path extension (often .git) matches the repository's specifier, as
different git clients inconsistently preserve the path extension of the
remote. Moving to the CanonicalPackageURL also ensures that absolute
paths are treated as urls with a file:// scheme, matching git's
behavior.