Add `async`-friendly `WritableStream`, create `_AsyncFileSystem` module (#7831)
Existing `WritableByteStream` protocol used by `AsyncProcess` is not
thread-safe. It's new `WritableStream` replacement is, with an added
conformance to the existing `OpenWritableFile`. That one is built on top
`AsyncFS`, which has been moved to the `Basics` module. That allows us
to adopt it in `AsyncProcess` in the future.

This is NFC, new API is adopted solely by `QueryEngine`, which is
exercised only in tests.