[Build] BuildPlan: Always discover test modules through their aggrega… (#7879)
…te products

### Motivation:

Fixes a situation when tests were build for both host and target even
though they should have been built only for the host (i.e. when one of
the modules depends on a macro).

### Modifications:

- Adjusted `BuildPlan.computeDestinations` to never add `test` modules
to successor list of a package and allow `test` products to always
iterate their module dependencies instead.

### Result:

If one of the test targets depends on a macro we'd never attempt to
build others for "target".