[WIP] Conditionally enable or disable XCTest/Swift Testing based on `swift build` arguments.
This PR conditionally enables/disables the library-specific content added to a
test product's synthesized main function based on what flags are passed to
`swift build`. The exact effects of these flags are currently undocumented. This
change will allow Foundation to explicitly opt out of linking to Swift Testing
(which links to Foundation, producing a circular reference.)