Fix `InternalImportsByDefault` diagnostics in _AsyncFileSystem.
Both the `StrictConcurrency` and `InternalImportsByDefault` features became
"upcoming" features instead of "experimental" features in the Swift 6 compiler.
This means that specifying those features with `.enableExperimentalFeature()`
no longer works with newer toolchains, and therefore these settings in
SwiftPM's Package.swift have become inactive. I'm making the compiler more
lenient, so that `-enable-experimental-feature` will enable the corresponding
upcoming feature if relevant (https://github.com/swiftlang/swift/pull/75962)
but in the meantime it appears that some of the code in SwiftPM needs some
fixes to continue to be accepted by the compiler with these features on.