Fix discrepancies with usage of package display name/identity
swift package edit was reliant on the user-passed packageName argument
to be able to both find the dependency (which needs the identity) as
well as to match the manifest display name.

Default to treating the user-passed argument as a package identity,
and fix the comparisons made to the manifest by instead assuring
that the canonical locations for the manifest and the dependency
match. The display name was intended to be deprecated.