[Tests] NFC: Remove overload of `GeneralizedTime.init(_:)`
`swift-certificates` now implements initializer that accepts
`Date` and doesn't throw which means this method is redundant
but also incorrect because overloading only on `throws` is
ambiguous in most cases due to throws checking not being part
of the expression checking.

Uses of this overload type-checked only because it was always
preferred before due to old type-checker hacks that we are
trying to remove.