Provide a list of executables in package description
The `swift run` subcommand runs executables that come from a
variety of places. Typically, users are expecting to be able to
execute products and targets for their package, iterating
quickly with their code changes. You can also run executable
products from packages that are part of the dependency list. It
turns out that all of the executable products of the root
package's transitive closure are available.

Discovery of the available executables is limited to either direct
inspection of the Package.swift, and the `swift package describe`.
However, these only work to find the executables of the current
package, not in the dependent packages. This has real implications
for a package maintainer since duplicate executable names are
forbidden among the closure of the package's dependencies.

Provide the ability to see the list of available executables for
a package and its dependencies via the `swift package describe`
subcommand. Qualify executables that aren't in the current package
with their package of origin for traceability purposes. Since the
description can be formatted as JSON this information can be used
by third-party tools.