Add a new show-executables package subcommand for all executables
The `swift run` subcommand runs executables that come from a variety of places. Typically, users
are expecting to be able to execute products and targets for their package, iterating quickly with
their code changes. You can also run executable products from packages that are part of the
dependency list. It turns out that all of the executable products of the root package's transitive
closure are available.

Discovery of the available executables is limited to either direct inspection of the Package.swift,
the `swift package describe`, and `swift package completion-tool list-executables`. However, these
only work to find the executables of the current package, not in the dependent packages. This has
real implications for a package maintainer since duplicate executable names are forbidden among the
closure of the package's dependencies.

Provide the ability to show the complete list of available executables from the package with a new
`swift package show-executables` subcommand. Qualify executables that aren't in the current package
with their package of origin for traceability purposes. The list can be formatted as JSON so that
this information can be used by third-party tools.