Reenable testUsage() as testInvalidUsage()
I believe its expected for the argument `-halp` to show the help string
in swift-argument-parser, as if there is a `-h` or `--help` anywhere in
the arguments all other inputs are disregarded and the help string is
shown.

Instead use an invalid argument to drive the test and reenable it.