Documentation: Fix the typo in the word `Continuous` (#8027)
There was a recurring typo in the word `Continuous`.

### Motivation:

To improve the overall quality of the project.

### Modifications:

The typo appeared in a file name and one sentence, both of which have
been corrected.

### Result:

These changes are backward compatible as they pertain to documentation
only, with no impact on the code.